[unittest] split out Config and TestHarness into separate files, minimize imports
